+++ client/src/components/Trees.js
@@ -5,18 +5,16 @@ Trees is a contaier component (no visual representation of it's own and concerne
 handling the comms between the tree view components and the store/models)
 
 */
-import React from "react";
-import PropTypes from "prop-types";
-import { connect } from "react-redux";
+import React from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
 
-import TableToolbar from "./TableToolbar";
-import TreeTable from "./TreeTable";
+import TreeTable from './TreeTable';
 
 const Trees = props => ({
   render() {
     return (
       <div className="trees">
-        <TableToolbar />
         <TreeTable />
       </div>
     );
